+++ HoloPipelines/tests/test_lung_segmentation_pipeline.py
@@ -5,8 +5,8 @@ from unittest import mock
 import pytest
 
 from core.pipelines import lung_segmentation
-from .input_data import sample_medical_data
-from .shared_fixtures import (
+from tests.utils.input_data import sample_medical_data
+from tests.utils.shared_fixtures import (
     patch_jobs_io_and_create_dirs,
     mock_send_to_holostorage_accessor,
 )
